requirements.txt
SQLAlchemy
mysqlclient 
python-dotenv 
pandas
pytest